First pass of networking support
Basic URL fetching works on Linux and PS3. Compiles on Windows
but is not working until we can initialize Winsock.

Move object watcher from lbshell into base.

Modify gyp files to reference googleurl/googleurl.gyp rather than
build/temp_gyp/googleurl.gyp, and third_party/openssl.

Disable net_resources actions for Cobalt- we don't have tools/grit,
and I don't think net_resources is needed.

Set use_native_http_stack variable default to 0.

Some DIAL compilation fixes for Linux- don't use file-scope
std::strings as they require global destructors to run.

Fix a few compiler warnings in under MSVC related to
signed/unsigned conversion and type narrowing.

Add actual_target_arch to filename_rules.gypi, to allow
building lbshell/src/*_win.cc on Windows.

Fix a few lint errors.

+++ net/base/int128.h
@@ -110,7 +110,8 @@ inline uint128::uint128(const uint128 &v) : lo_(v.lo_), hi_(v.hi_) { }
 inline uint128::uint128(const uint128_pod &v) : lo_(v.lo), hi_(v.hi) { }
 inline uint128::uint128(uint64 bottom) : lo_(bottom), hi_(0) { }
 inline uint128::uint128(uint32 bottom) : lo_(bottom), hi_(0) { }
-inline uint128::uint128(int bottom) : lo_(bottom), hi_(0) {
+inline uint128::uint128(int bottom)
+    : lo_(static_cast<uint32>(bottom)), hi_(0) {
   if (bottom < 0) {
     --hi_;
   }
